extract addSortedSectionsAndPages
