+++ src/utils/hooks/useGetFullReviewStructure.tsx
@@ -11,6 +11,7 @@ import addElementsById from '../helpers/structure/addElementsById'
 import generateReviewProgress from '../helpers/structure/generateReviewProgress'
 import { cloneDeep } from '@apollo/client/utilities'
 import generateReviewSectionActions from '../helpers/structure/generateReviewSectionActions'
+import addSortedSectionsAndPages from '../helpers/structure/addSortedSectionsAndPages'
 
 interface useGetFullReviewStructureProps {
   fullApplicationStructure: FullStructure
@@ -89,19 +90,6 @@ const useGetFullReviewStructure = ({
     fullReviewStructure,
     error: error?.message,
   }
-}
-
-const addSortedSectionsAndPages = (newStructure: FullStructure): FullStructure => {
-  const sortedSections = Object.values(newStructure.sections).sort(
-    (sectionOne, sectionTwo) => sectionOne.details.index - sectionTwo.details.index
-  )
-  const sortedPages = sortedSections
-    .map((section) =>
-      Object.values(section.pages).sort((pageOne, pageTwo) => pageOne.number - pageTwo.number)
-    )
-    .flat()
-
-  return { ...newStructure, sortedPages, sortedSections }
 }
 
 const getFilteredSections = (sectionIds: number[], sections: SectionStateNEW[]) => {
